forgot password changes

+++ lib/models/user_habit_expense_category.dart
@@ -1,7 +1,7 @@
 class UserHabitExpenseCategory {
   String? expenseCatName;
   String? categoryColor;
-  int? habitCatPercentage;
+  double? habitCatPercentage;
 
   UserHabitExpenseCategory({
       this.expenseCatName,
